requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
python-dotenv
pypdf
python-docx
openpyxl
chromadb
sentence-transformers
